Added ServerView
It seems like this works correctly except none of the message types sent to the server view have a component to render them, so nothing shows up for now.

Somewhat fixed server item style

However we'll probably need a custom NavLink or something to make it so we can style the arrow, too.

Fixed raw message output

Fixed styling for server link in sidebar + refactored RouteStore