Refactor routing, passing apps
Refactored the routing so that we avoid cloning elements and double rendering in SquelchView, as well as avoiding the false prop validation errors.

Added wrapper components for each type of view so that the views themselves can be simplified and only be passed the props they need, instead of the whole state.

Added a convenience function for shallowly checking if certain props have changed. Shortens the shouldComponentUpdate code a bit for some components.

Resolves #37
Resolves #50